+++ extensions/amp-form/0.1/validation-bubble.js
@@ -32,7 +32,7 @@ export class ValidationBubble {
     /** @private @const {string} */
     this.id_ = id;
 
-    /** @private @const {!../../../src/service/viewport-impl.Viewport} */
+    /** @private @const {!../../../src/service/viewport/viewport-impl.Viewport} */
     this.viewport_ = Services.viewportForDoc(ampdoc);
 
     /** @private @const {!../../../src/service/vsync-impl.Vsync} */
