+++ src/pull-to-refresh.js
@@ -41,7 +41,7 @@ export function installPullToRefreshBlocker(win) {
 export class PullToRefreshBlocker {
   /**
    * @param {!Document} doc
-   * @param {!./service/viewport-impl.Viewport} viewport
+   * @param {!./service/viewport/viewport-impl.Viewport} viewport
    */
   constructor(doc, viewport) {
     /** @private {!Document} */
